Moved pyparloop in parloop.py
